# AstGen: use string index as key for string table