@andrewrk 2021-09-07
Here are some timings on my laptop:

## master branch

test2.zig is just `pub fn main() void {}` but it is an appropriate test because if you wipe the cache beforehand all of std will get AstGen'd.

```
[nix-shell:~/dev/zig/build-release]$ perf stat -e cache-misses -e cache-references -e instructions -e cycles -e branches -e branch-misses ./zig-out/bin/zig build-exe test2.zig

 Performance counter stats for './zig-out/bin/zig build-exe test2.zig':

         2,842,425      cache-misses              #    9.591 % of all cache refs    
        29,636,102      cache-references                                            
     1,579,654,888      instructions              #    1.18  insn per cycle         
     1,339,267,718      cycles                                                      
       267,919,133      branches                                                    
         6,796,142      branch-misses             #    2.54% of all branches        

       0.034473438 seconds time elapsed

       0.327205000 seconds user
       0.047934000 seconds sys



[nix-shell:~/dev/zig/build-release]$ $(which time) -v ./zig-out/bin/zig build-exe test2.zig
	Command being timed: "./zig-out/bin/zig build-exe test2.zig"
	User time (seconds): 0.30
	System time (seconds): 0.05
	Percent of CPU this job got: 1041%
	Elapsed (wall clock) time (h:mm:ss or m:ss): 0:00.03
	Average shared text size (kbytes): 0
	Average unshared data size (kbytes): 0
	Average stack size (kbytes): 0
	Average total size (kbytes): 0
	Maximum resident set size (kbytes): 88200
	Average resident set size (kbytes): 0
	Major (requiring I/O) page faults: 0
	Minor (reclaiming a frame) page faults: 13214
	Voluntary context switches: 119
	Involuntary context switches: 20
	Swaps: 0
	File system inputs: 0
	File system outputs: 35680
	Socket messages sent: 0
	Socket messages received: 0
	Signals delivered: 0
	Page size (bytes): 4096
	Exit status: 0
```

## This PR

```
[nix-shell:~/Downloads/zig/build-release]$ perf stat -e cache-misses -e cache-references -e instructions -e cycles -e branches -e branch-misses ./zig-out/bin/zig build-exe test2.zig

 Performance counter stats for './zig-out/bin/zig build-exe test2.zig':

         2,759,377      cache-misses              #    9.700 % of all cache refs    
        28,446,010      cache-references                                            
     1,581,242,237      instructions              #    1.20  insn per cycle         
     1,319,606,058      cycles                                                      
       271,393,367      branches                                                    
         6,953,320      branch-misses             #    2.56% of all branches        

       0.033078899 seconds time elapsed

       0.289673000 seconds user
       0.057523000 seconds sys



[nix-shell:~/Downloads/zig/build-release]$ $(which time) -v ./zig-out/bin/zig build-exe test2.zig
	Command being timed: "./zig-out/bin/zig build-exe test2.zig"
	User time (seconds): 0.31
	System time (seconds): 0.04
	Percent of CPU this job got: 1017%
	Elapsed (wall clock) time (h:mm:ss or m:ss): 0:00.03
	Average shared text size (kbytes): 0
	Average unshared data size (kbytes): 0
	Average stack size (kbytes): 0
	Average total size (kbytes): 0
	Maximum resident set size (kbytes): 88212
	Average resident set size (kbytes): 0
	Major (requiring I/O) page faults: 0
	Minor (reclaiming a frame) page faults: 13150
	Voluntary context switches: 157
	Involuntary context switches: 21
	Swaps: 0
	File system inputs: 0
	File system outputs: 35688
	Socket messages sent: 0
	Socket messages received: 0
	Signals delivered: 0
	Page size (bytes): 4096
	Exit status: 0
```

## Summary

 * wall clock time improved by 4%
 * Cache-misses improved by 3%
 * Peak memory, instruction count: no difference.
 * Branches worsened by 1%